api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai>=1.0.0
python-dotenv==0.19.0
PyMuPDF==1.23.7
Pillow==10.1.0